# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fine-Tune Labs API
  version: 1.0.0
  description: |

    # Fine-Tune Labs Public API

    Production-grade API for LLMOps, model training, and monitoring.

    ## Features

    - 🎯 **Widget Apps** - Deploy production monitoring widgets
    - 📊 **Analytics & Metrics** - Real-time model performance tracking
    - 🚀 **Inference Deployment** - Deploy and manage model endpoints
    - 🤖 **Model Management** - CRUD operations for LLM models
    - 📈 **Training Jobs** - Monitor fine-tuning progress

    ## Authentication

    All endpoints require Bearer token authentication:

    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```

    Get your API token from the Fine-Tune Labs dashboard.

    ## Base URL

    - **Development**: `http://localhost:3000`
    - **Production**: `https://finetunelab.ai`

    ## Rate Limits

    - 1000 requests per minute per API key
    - 10,000 requests per day per API key

    ## Support

    - Documentation: https://finetunelab.ai/docs
    - Email: support@finetunelab.ai
  contact:
    name: Fine-Tune Labs API Support
    email: support@finetunelab.ai
    url: https://finetunelab.ai/support
  license:
    name: Proprietary
    url: https://finetunelab.ai/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://finetunelab.ai
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        code:
          type: string
          description: Error code
      required:
        - error
...
    WidgetApp:
      type: object
      properties:
        id:
          type: string
          description: Unique widget app ID
          example: widget_abc123
        user_id:
          type: string
          description: User ID who owns this widget
        name:
          type: string
          description: Widget app name
          example: Customer Support Widget
        description:
          type: string
          description: Widget description
          example: Production chatbot for customer support
        api_key:
          type: string
          description: API key for this widget
          example: wgt_key_xyz789abc
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
security:
  - bearerAuth: []
...
paths:
  /api/widget-apps:
    get:
      summary: List widget apps
      description: Retrieve all widget apps for the authenticated user
      tags:
        - Widget Apps
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved widget apps
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 2
                  apps:
                    type: array
                    items:
                      $ref: '#/components/schemas/WidgetApp'
        '401':
          description: Unauthorized - Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create widget app
      description: Create a new widget app for production monitoring. Returns an API token (shown only once).
      tags:
        - Widget Apps
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Widget app name
                  example: My Production App
                description:
                  type: string
                  description: Optional description
                  example: Monitors chatbot performance
                allowed_origins:
                  type: array
                  items:
                    type: string
                  description: CORS allowed origins
                  example:
                    - https://myapp.com
                    - https://www.myapp.com
      responses:
        '201':
          description: Widget app created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  app:
                    type: object
                    properties:
                      id:
                        type: string
                        example: app_abc123xyz
                      token:
                        type: string
                        description: API token (only shown once, save it!)
                        example: widget_xyz789abc...
                      name:
                        type: string
                        example: My Production App
                      description:
                        type: string
                        example: Monitors chatbot performance
                      allowed_origins:
                        type: array
                        items:
                          type: string
                        example:
                          - https://myapp.com
        '400':
          description: Bad request - Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...